backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
requests>=2.31
psycopg2-binary>=2.9
firebase-admin>=6.4
